DROID-3965 ui
