+++ app/src/main/java/com/anytypeio/anytype/ui/widgets/types/LinkWidget.kt
@@ -1,14 +1,11 @@
 package com.anytypeio.anytype.ui.widgets.types
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -23,12 +20,13 @@ import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.hapticfeedback.HapticFeedbackType
 import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.anytypeio.anytype.R
+import com.anytypeio.anytype.core_models.Id
 import com.anytypeio.anytype.core_ui.foundation.noRippleClickable
 import com.anytypeio.anytype.core_ui.views.HeadlineSubheading
+import com.anytypeio.anytype.core_ui.widgets.ListWidgetObjectIcon
 import com.anytypeio.anytype.presentation.widgets.DropDownMenuAction
 import com.anytypeio.anytype.presentation.widgets.Widget
 import com.anytypeio.anytype.presentation.widgets.WidgetId
@@ -43,7 +41,7 @@ fun LinkWidgetCard(
     onDropDownMenuAction: (DropDownMenuAction) -> Unit,
     isInEditMode: Boolean,
     hasReadOnlyAccess: Boolean = false,
-    onWidgetMenuTriggered: (WidgetId) -> Unit,
+    onObjectCheckboxClicked: (Id, Boolean) -> Unit
 ) {
     val isCardMenuExpanded = remember {
         mutableStateOf(false)
@@ -62,14 +60,7 @@ fun LinkWidgetCard(
                 color = colorResource(id = R.color.dashboard_card_background)
             )
             .then(
-                if (isInEditMode) {
-                    Modifier.noRippleClickable {
-                        isCardMenuExpanded.value = !isCardMenuExpanded.value
-                        if (isCardMenuExpanded.value == true) {
-                            onWidgetMenuTriggered(item.id)
-                        }
-                    }
-                } else if (hasReadOnlyAccess) {
+                if (hasReadOnlyAccess) {
                     Modifier.noRippleClickable {
                         onWidgetSourceClicked(item.id, item.source)
                     }
@@ -86,50 +77,35 @@ fun LinkWidgetCard(
                 }
             )
     ) {
-        Box(
+
+        Row (
             Modifier
-                .padding(vertical = 6.dp)
                 .fillMaxWidth()
-                .height(40.dp)
+                .height(52.dp)
+                .padding(horizontal = 16.dp),
+            verticalAlignment = Alignment.CenterVertically
         ) {
+
+            ListWidgetObjectIcon(
+                iconSize = 18.dp,
+                icon = item.icon,
+                modifier = Modifier.padding(end = 12.dp),
+                onTaskIconClicked = { isChecked ->
+                    onObjectCheckboxClicked(
+                        item.source.id,
+                        isChecked
+                    )
+                }
+            )
 
             Text(
                 text = item.getPrettyName(),
                 maxLines = 1,
                 overflow = TextOverflow.Ellipsis,
-                modifier = Modifier
-                    .align(Alignment.CenterStart)
-                    .padding(
-                        start = 16.dp,
-                        end = if (isInEditMode) 76.dp else 32.dp
-                    ),
+                modifier = Modifier.weight(1f),
                 style = HeadlineSubheading,
                 color = colorResource(id = R.color.text_primary),
             )
-            AnimatedVisibility(
-                visible = isInEditMode,
-                modifier = Modifier
-                    .align(Alignment.CenterEnd)
-                    .padding(end = 48.dp),
-                enter = fadeIn(),
-                exit = fadeOut()
-            ) {
-                Box {
-                    Image(
-                        painterResource(R.drawable.ic_widget_three_dots),
-                        contentDescription = "Widget menu icon",
-                        modifier = Modifier
-                            .noRippleClickable {
-                                isHeaderMenuExpanded.value = !isHeaderMenuExpanded.value
-                            }
-                    )
-                    WidgetMenu(
-                        isExpanded = isHeaderMenuExpanded,
-                        onDropDownMenuAction = onDropDownMenuAction,
-                        canEditWidgets = !isInEditMode
-                    )
-                }
-            }
         }
         WidgetMenu(
             isExpanded = isCardMenuExpanded,
